requirements.txt
# Core Flask and Google Cloud
Flask==3.0.3
Flask-CORS==4.0.0
Flask-Limiter==3.3.1
marshmallow==3.19.0
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.56.0
vertexai
gunicorn

# Firebase Authentication
firebase-admin==6.5.0
Pyrebase4==4.8.0

# Helper Libraries
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1

# Google Cloud KMS for encryption
google-cloud-kms==2.20.0